# append faster-coco-eval as base backend for COCODetection