--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -29,7 +29,7 @@ jobs:
         type: string
       paddle:
         type: string
-        default: "2.2.2"
+        default: "2.3.2"
       tensorflow:
         type: string
         default: "2.4"