--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -108,96 +108,103 @@ jobs:
           command: |
             mpirun -np 2 pytest -vv --capture=no --with-mpi tests/test_core/test_dist_backends
 
-  build_cu112:
+  build_cuda:
+    parameters:
+      torch:
+        type: string
+      cuda:
+        type: enum
+        enum: ["11.7", "11.8"]
+      cudnn:
+        type: integer
+        default: 8
     machine:
-      image: ubuntu-2004-cuda-11.2:202103-01
-    resource_class: gpu.nvidia.small
+      image: linux-cuda-11:default
+      # docker_layer_caching: true
+    resource_class: gpu.nvidia.small.multi
     steps:
       - checkout
       - run:
-          # CUDNN is required for Paddle to use GPU, otherwise, the test cases will report an error.
-          name: Install CUDNN
+          # CLoning repos in VM since Docker doesn't have access to the private key
+          name: Clone Repos
           command: |
-            OS=ubuntu2004
-            cudnn_version=8.1.0.77
-            cuda_version=cuda11.2
-            wget https://developer.download.nvidia.com/compute/cuda/repos/${OS}/x86_64/cuda-${OS}.pin
-            sudo mv cuda-${OS}.pin /etc/apt/preferences.d/cuda-repository-pin-600
-            sudo apt-key adv --fetch-keys https://developer.download.nvidia.com/compute/cuda/repos/${OS}/x86_64/3bf863cc.pub
-            sudo add-apt-repository "deb https://developer.download.nvidia.com/compute/cuda/repos/${OS}/x86_64/ /"
+            git clone -b main --depth 1 ssh://git@github.com/open-mmlab/mmengine.git /home/circleci/mmengine
+      - run:
+          name: Install nvidia-container-toolkit and Restart Docker
+          command: |
             sudo apt-get update
-            sudo apt-get install libcudnn8=${cudnn_version}-1+${cuda_version}
-            sudo apt-get install libcudnn8-dev=${cudnn_version}-1+${cuda_version}
+            sudo apt-get install -y nvidia-container-toolkit
+            sudo systemctl restart docker
       - run:
-          name: Install PyTorch, Paddle and OneFlow via pip
+          name: Build Docker image
           command: |
-            pyenv global 3.9.2
-            pip install --upgrade pip
-            python -V
-            pip install torch==1.7.1+cu110 -f https://download.pytorch.org/whl/torch_stable.html
-            pip install paddlepaddle-gpu==2.3.2.post112 -f https://www.paddlepaddle.org.cn/whl/linux/mkl/avx/stable.html
-            # Install Oneflow often failed, we just skip it if failed.
-            pip install --pre oneflow -f https://staging.oneflow.info/branch/master/cu112 "numpy<1.24.0" || true
+            docker build .circleci/docker -t mmeval:gpu --build-arg PYTORCH=<< parameters.torch >> --build-arg CUDA=<< parameters.cuda >> --build-arg CUDNN=<< parameters.cudnn >>
+            docker run --gpus all -t -d -v /home/circleci/project:/mmeval -v /home/circleci/mmengine:/mmengine -w /mmeval --name mmeval mmeval:gpu
+            docker exec mmeval apt-get install -y git
       - run:
-          name: Install mmeval and dependencies
+          name: Install mmeval dependencies
           command: |
-            pip install -r requirements/tests.txt
-            pip install -r requirements/optional.txt
-            pip install -e .
+            docker exec mmeval pip install paddlepaddle-gpu==2.3.2.post112 -f https://www.paddlepaddle.org.cn/whl/linux/mkl/avx/stable.html
+            docker exec mmeval pip install -r requirements/tests.txt
+            docker exec mmeval pip install -r requirements/optional.txt
+      - run:
+          name: Build and install
+          command: |
+            docker exec mmeval pip install -e .
       - run:
           name: Run unittests
           command: |
-            pytest -vv tests/
-
+            docker exec mmeval python -m pytest -vv tests/
 
 workflows:
   unit_tests:
     jobs:
       - lint
-      - build_cpu:
-          name: build_cpu_torch1.6_tf2.4
-          torch: 1.6.0
-          tensorflow: "2.4"
-          requires:
-            - lint
-      - build_cpu:
-          name: build_cpu_torch1.7_tf_2.6
-          torch: 1.7.0
-          tensorflow: "2.6"
-          requires:
-            - lint
-      - build_cpu:
-          name: build_cpu_torch1.8_tf2.7_py3.9
-          torch: 1.8.0
-          python: "3.9.0"
-          tensorflow: "2.7"
-          requires:
-            - lint
-      - build_cpu:
-          name: build_cpu_torch1.9_paddle2.3_tf2.8_py3.8
-          torch: 1.9.0
-          paddle: 2.3.2
-          tensorflow: "2.8"
-          python: "3.8.12"
-          requires:
-            - lint
-      - build_cpu:
-          name: build_cpu_torch1.9_paddle2.10_py3.9
-          torch: 1.9.0
-          paddle: 2.4.0rc0
-          tensorflow: "2.10"
-          python: "3.9.0"
-          requires:
-            - lint
-      - build_mpirun_and_tf:
-          name: build_mpirun_and_tf_2.7
-          tensorflow: "2.7"
-          requires:
-            - build_cpu_torch1.6_tf2.4
-            - build_cpu_torch1.7_tf_2.6
-            - build_cpu_torch1.8_tf2.7_py3.9
-            - build_cpu_torch1.9_paddle2.3_tf2.8_py3.8
-            - build_cpu_torch1.9_paddle2.10_py3.9
-      - build_cu112:
-          requires:
-            - build_mpirun_and_tf_2.7
+      # - build_cpu:
+      #     name: build_cpu_torch1.6_tf2.4
+      #     torch: 1.6.0
+      #     tensorflow: "2.4"
+      #     requires:
+      #       - lint
+      # - build_cpu:
+      #     name: build_cpu_torch1.7_tf_2.6
+      #     torch: 1.7.0
+      #     tensorflow: "2.6"
+      #     requires:
+      #       - lint
+      # - build_cpu:
+      #     name: build_cpu_torch1.8_tf2.7_py3.9
+      #     torch: 1.8.0
+      #     python: "3.9.0"
+      #     tensorflow: "2.7"
+      #     requires:
+      #       - lint
+      # - build_cpu:
+      #     name: build_cpu_torch1.9_paddle2.3_tf2.8_py3.8
+      #     torch: 1.9.0
+      #     paddle: 2.3.2
+      #     tensorflow: "2.8"
+      #     python: "3.8.12"
+      #     requires:
+      #       - lint
+      # - build_cpu:
+      #     name: build_cpu_torch1.9_paddle2.10_py3.9
+      #     torch: 1.9.0
+      #     paddle: 2.4.0rc0
+      #     tensorflow: "2.10"
+      #     python: "3.9.0"
+      #     requires:
+      #       - lint
+      # - build_mpirun_and_tf:
+      #     name: build_mpirun_and_tf_2.7
+      #     tensorflow: "2.7"
+      #     requires:
+      #       - build_cpu_torch1.6_tf2.4
+      #       - build_cpu_torch1.7_tf_2.6
+      #       - build_cpu_torch1.8_tf2.7_py3.9
+      #       - build_cpu_torch1.9_paddle2.3_tf2.8_py3.8
+      #       - build_cpu_torch1.9_paddle2.10_py3.9
+      - build_cuda:
+          name: build_cuda117_torch1.8_paddle2.3
+          torch: 1.8.1
+          cuda: "11.7"