--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -160,50 +160,50 @@ workflows:
   unit_tests:
     jobs:
       - lint
-      # - build_cpu:
-      #     name: build_cpu_torch1.6_tf2.4
-      #     torch: 1.6.0
-      #     tensorflow: "2.4"
-      #     requires:
-      #       - lint
-      # - build_cpu:
-      #     name: build_cpu_torch1.7_tf_2.6
-      #     torch: 1.7.0
-      #     tensorflow: "2.6"
-      #     requires:
-      #       - lint
-      # - build_cpu:
-      #     name: build_cpu_torch1.8_tf2.7_py3.9
-      #     torch: 1.8.0
-      #     python: "3.9.0"
-      #     tensorflow: "2.7"
-      #     requires:
-      #       - lint
-      # - build_cpu:
-      #     name: build_cpu_torch1.9_paddle2.3_tf2.8_py3.8
-      #     torch: 1.9.0
-      #     paddle: 2.3.2
-      #     tensorflow: "2.8"
-      #     python: "3.8.12"
-      #     requires:
-      #       - lint
-      # - build_cpu:
-      #     name: build_cpu_torch1.9_paddle2.10_py3.9
-      #     torch: 1.9.0
-      #     paddle: 2.4.0rc0
-      #     tensorflow: "2.10"
-      #     python: "3.9.0"
-      #     requires:
-      #       - lint
-      # - build_mpirun_and_tf:
-      #     name: build_mpirun_and_tf_2.7
-      #     tensorflow: "2.7"
-      #     requires:
-      #       - build_cpu_torch1.6_tf2.4
-      #       - build_cpu_torch1.7_tf_2.6
-      #       - build_cpu_torch1.8_tf2.7_py3.9
-      #       - build_cpu_torch1.9_paddle2.3_tf2.8_py3.8
-      #       - build_cpu_torch1.9_paddle2.10_py3.9
+      - build_cpu:
+          name: build_cpu_torch1.6_tf2.4
+          torch: 1.6.0
+          tensorflow: "2.4"
+          requires:
+            - lint
+      - build_cpu:
+          name: build_cpu_torch1.7_tf_2.6
+          torch: 1.7.0
+          tensorflow: "2.6"
+          requires:
+            - lint
+      - build_cpu:
+          name: build_cpu_torch1.8_tf2.7_py3.9
+          torch: 1.8.0
+          python: "3.9.0"
+          tensorflow: "2.7"
+          requires:
+            - lint
+      - build_cpu:
+          name: build_cpu_torch1.9_paddle2.3_tf2.8_py3.8
+          torch: 1.9.0
+          paddle: 2.3.2
+          tensorflow: "2.8"
+          python: "3.8.12"
+          requires:
+            - lint
+      - build_cpu:
+          name: build_cpu_torch1.9_paddle2.10_py3.9
+          torch: 1.9.0
+          paddle: 2.4.0rc0
+          tensorflow: "2.10"
+          python: "3.9.0"
+          requires:
+            - lint
+      - build_mpirun_and_tf:
+          name: build_mpirun_and_tf_2.7
+          tensorflow: "2.7"
+          requires:
+            - build_cpu_torch1.6_tf2.4
+            - build_cpu_torch1.7_tf_2.6
+            - build_cpu_torch1.8_tf2.7_py3.9
+            - build_cpu_torch1.9_paddle2.3_tf2.8_py3.8
+            - build_cpu_torch1.9_paddle2.10_py3.9
       - build_cuda:
           name: build_cuda117_torch1.8_paddle2.3
           torch: 1.8.1