--- a/requirements/optional.txt
+++ b/requirements/optional.txt
@@ -1,2 +1,3 @@
 pycocotools
+faster-coco-eval
 shapely