--- a/requirements/optional.txt
+++ b/requirements/optional.txt
@@ -1,3 +1,3 @@
-pycocotools
 faster-coco-eval
+pycocotools
 shapely