--- a/requirements/optional.txt
+++ b/requirements/optional.txt
@@ -1,3 +1,3 @@
+faster-coco-eval @ git+https://github.com/MiXaiLL76/faster_coco_eval@dev
 pycocotools
 shapely
-faster-coco-eval @ git+https://github.com/MiXaiLL76/faster_coco_eval@dev