--- a/requirements/runtime.txt
+++ b/requirements/runtime.txt
@@ -1,4 +1,4 @@
-numpy
+numpy>1.20.1
 plum-dispatch<2.0.0
 pyyaml
 rich