<https://github.com/MiXaiLL76/faster_coco_eval>

## Motivation

I often use this project, but I saw it abandoned and without a public repository on github. Also, part of the project remained unfinished for a long time. I implemented some of the author's ideas and decided to make the results publicly available.

## Modification

This implementation greatly speeds up the evaluation time for coco's AP metrics, especially when dealing with a high number of instances in an image.

## Use cases

```py
COCODetection(..., faster=True)
```

| Type | COCOeval    | COCOeval_faster | Profit      |
| ---- | ----------- | --------------- | ----------- |
| bbox | 18.477 sec. | 7.345 sec.      | 2.5x faster |
| segm | 29.819 sec. | 15.840 sec.     | 2x faster   |