@MiXaiLL76 2024-02-01
@zhouzaida @C1rN09 
cannot run last test with GPU
```
Job was rejected because resource class gpu.nvidia.small, image ubuntu-2004-cuda-11.2:202103-01 is not a valid resource class
```

mb 
<https://discuss.circleci.com/t/linux-cuda-deprecation-and-image-policy/48568>